create

+++ src/main/java/com/spring/boot/sensor/service/impl/AdminServiceImpl.java
@@ -213,11 +213,15 @@ public class AdminServiceImpl implements AdminService {
         if (parameterM.getDelete() == 1) {
             if (StringUtils.isBlank(parameterM.getIds())) return ResultUtil.errorWithMessage("请先选择要操作的数据");
             if (parameterM.getIds().split(",").length > 1) return ResultUtil.errorWithMessage("只能选择一条数据");
+            Role role = roleMapper.findById(Integer.parseInt(parameterM.getIds()));
+            if (role.getIslock() == 1) return ResultUtil.errorWithMessage("不能删除内置角色");
             roleMapper.deleteById(Integer.parseInt(parameterM.getIds()));
             return ResultUtil.ok();
         } else if (parameterM.getDelete() == 2) {
             if (StringUtils.isBlank(parameterM.getIds())) return ResultUtil.errorWithMessage("请先选择要操作的数据");
             for (String id : parameterM.getIds().split(",")) {
+                Role role = roleMapper.findById(Integer.parseInt(id));
+                if (role.getIslock() == 1) return ResultUtil.errorWithMessage("不能删除内置角色");
                 roleMapper.deleteById(Integer.parseInt(id));
             }
             return ResultUtil.ok();
